+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## AppLovin Android Mediation Adapter Changelog
 
+#### Next Version
+- Adapter can now call onUserEarnedReward on videoPlaybackEnded callback.
+
 #### Version 13.1.0.1
 - Updated the minimum required Android API level to 23.
 - Updated the minimum required Google Mobile Ads SDK version to 24.0.0.